Add support for more AppInsightTypes
- Moves the related tests to a new test file: NLFPluginAppInsightTest
